+++ .eslintrc.js
@@ -1,6 +1,6 @@
 module.exports = {
   extends: "divid",
   parserOptions: {
-    project: ["./src/tsconfig.json", "./test/tsconfig.json"]
+    project: ["./src/tsconfig.json"]
   }
 };